Adapt to BoCA library changes.

+++ components/faac/faac.h
@@ -39,7 +39,7 @@ namespace BoCA
 
 			Int			 totalSamples;
 
-			Buffer<int32_t>		 samplesBuffer;
+			Buffer<int16_t>		 samplesBuffer;
 
 			Int			 EncodeFrames(Bool);
 			Int			 ProcessPackets(const Buffer<unsigned char> &, const Array<Int> &, Bool);
